+++ rootdriver/utils/__init__.py
@@ -1,6 +1,6 @@
 from .time import get_iso_timestamp
 from .build_message import build_user_message, build_tool_message, build_system_message, build_assistant_message, build_message
-
+from .file import ensure_file_exist
 
 __all__ = [
     "get_iso_timestamp",
@@ -8,5 +8,6 @@ __all__ = [
     "build_user_message",
     "build_assistant_message",
     "build_tool_message",
-    "build_system_message"
+    "build_system_message",
+    "ensure_file_exist"
 ]